Remove TypeJsonUtils
TypeJsonUtils are used to generate the stack representation for
complex types back to the ages there is no native block representation
for complex types in Presto.

With native block representation in Presto and the rewriting of
JSON cast, all methods in TypeJsonUtils are no longer used except
appendToBlockBuilder, which are used in test code.

In this commit:
- Refactor TypeJsonUtils.appendToBlockBuilder() into StructuralTestUtil.
- Inline the usage of appendToBlockBuilder() in TestLearnAggregations.
- Remove TypeJsonUtils.
