--- presto-main/src/main/java/com/facebook/presto/type/TypeJsonUtils.java
@@ -1,100 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.type;
-
-import com.facebook.presto.spi.block.BlockBuilder;
-import com.facebook.presto.spi.type.Decimals;
-import com.facebook.presto.spi.type.SqlDecimal;
-import com.facebook.presto.spi.type.StandardTypes;
-import com.facebook.presto.spi.type.Type;
-import com.google.common.annotations.VisibleForTesting;
-import io.airlift.slice.Slice;
-import io.airlift.slice.Slices;
-
-import java.util.Map;
-
-import static com.facebook.presto.spi.type.RealType.REAL;
-import static java.lang.Float.floatToRawIntBits;
-
-public final class TypeJsonUtils
-{
-    private TypeJsonUtils() {}
-
-    @VisibleForTesting
-    public static void appendToBlockBuilder(Type type, Object element, BlockBuilder blockBuilder)
-    {
-        Class<?> javaType = type.getJavaType();
-        if (element == null) {
-            blockBuilder.appendNull();
-        }
-        else if (type.getTypeSignature().getBase().equals(StandardTypes.ARRAY) && element instanceof Iterable<?>) {
-            BlockBuilder subBlockBuilder = blockBuilder.beginBlockEntry();
-            for (Object subElement : (Iterable<?>) element) {
-                appendToBlockBuilder(type.getTypeParameters().get(0), subElement, subBlockBuilder);
-            }
-            blockBuilder.closeEntry();
-        }
-        else if (type.getTypeSignature().getBase().equals(StandardTypes.ROW) && element instanceof Iterable<?>) {
-            BlockBuilder subBlockBuilder = blockBuilder.beginBlockEntry();
-            int field = 0;
-            for (Object subElement : (Iterable<?>) element) {
-                appendToBlockBuilder(type.getTypeParameters().get(field), subElement, subBlockBuilder);
-                field++;
-            }
-            blockBuilder.closeEntry();
-        }
-        else if (type.getTypeSignature().getBase().equals(StandardTypes.MAP) && element instanceof Map<?, ?>) {
-            BlockBuilder subBlockBuilder = blockBuilder.beginBlockEntry();
-            for (Map.Entry<?, ?> entry : ((Map<?, ?>) element).entrySet()) {
-                appendToBlockBuilder(type.getTypeParameters().get(0), entry.getKey(), subBlockBuilder);
-                appendToBlockBuilder(type.getTypeParameters().get(1), entry.getValue(), subBlockBuilder);
-            }
-            blockBuilder.closeEntry();
-        }
-        else if (javaType == boolean.class) {
-            type.writeBoolean(blockBuilder, (Boolean) element);
-        }
-        else if (javaType == long.class) {
-            if (element instanceof SqlDecimal) {
-                type.writeLong(blockBuilder, ((SqlDecimal) element).getUnscaledValue().longValue());
-            }
-            else if (REAL.equals(type)) {
-                type.writeLong(blockBuilder, floatToRawIntBits(((Number) element).floatValue()));
-            }
-            else {
-                type.writeLong(blockBuilder, ((Number) element).longValue());
-            }
-        }
-        else if (javaType == double.class) {
-            type.writeDouble(blockBuilder, ((Number) element).doubleValue());
-        }
-        else if (javaType == Slice.class) {
-            if (element instanceof String) {
-                type.writeSlice(blockBuilder, Slices.utf8Slice(element.toString()));
-            }
-            else if (element instanceof byte[]) {
-                type.writeSlice(blockBuilder, Slices.wrappedBuffer((byte[]) element));
-            }
-            else if (element instanceof SqlDecimal) {
-                type.writeSlice(blockBuilder, Decimals.encodeUnscaledValue(((SqlDecimal) element).getUnscaledValue()));
-            }
-            else {
-                type.writeSlice(blockBuilder, (Slice) element);
-            }
-        }
-        else {
-            type.writeObject(blockBuilder, element);
-        }
-    }
-}
