+++ presto-ml/src/test/java/com/facebook/presto/ml/TestLearnAggregations.java
@@ -48,7 +48,6 @@ import java.util.Random;
 
 import static com.facebook.presto.spi.type.TypeSignature.parseTypeSignature;
 import static com.facebook.presto.testing.AggregationTestUtils.generateInternalAggregationFunction;
-import static com.facebook.presto.type.TypeJsonUtils.appendToBlockBuilder;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
 
@@ -112,17 +111,14 @@ public class TestLearnAggregations
         Random rand = new Random(0);
         for (int i = 0; i < datapoints; i++) {
             long label = rand.nextDouble() < 0.5 ? 0 : 1;
-            builder.row(label, mapSliceOf(BigintType.BIGINT, DoubleType.DOUBLE, 0, label + rand.nextGaussian()), "C=1");
+
+            BlockBuilder mapSliceBlockBuilder = new InterleavedBlockBuilder(ImmutableList.of(BigintType.BIGINT, DoubleType.DOUBLE), new BlockBuilderStatus(), 2);
+            BigintType.BIGINT.writeLong(mapSliceBlockBuilder, 0L);
+            DoubleType.DOUBLE.writeDouble(mapSliceBlockBuilder, label + rand.nextGaussian());
+
+            builder.row(label, mapSliceBlockBuilder.build(), "C=1");
         }
 
         return builder.build();
-    }
-
-    private static Block mapSliceOf(Type keyType, Type valueType, Object key, Object value)
-    {
-        BlockBuilder blockBuilder = new InterleavedBlockBuilder(ImmutableList.of(keyType, valueType), new BlockBuilderStatus(), 100);
-        appendToBlockBuilder(keyType, key, blockBuilder);
-        appendToBlockBuilder(valueType, value, blockBuilder);
-        return blockBuilder.build();
     }
 }
