+++ presto-tests/src/main/java/com/facebook/presto/tests/StructuralTestUtil.java
@@ -34,7 +34,7 @@ import io.airlift.slice.Slice;
 import java.math.BigDecimal;
 import java.util.List;
 
-import static com.facebook.presto.type.TypeJsonUtils.appendToBlockBuilder;
+import static com.facebook.presto.util.StructuralTestUtil.appendToBlockBuilder;
 import static com.google.common.base.Preconditions.checkArgument;
 
 public final class StructuralTestUtil
